--- controllers/clientBehaviour.js
@@ -1,1 +1,0 @@
-const clientBehaviour=require('../models/clientBehaviour');
