--- controllers/containsController.js
@@ -1,1 +1,0 @@
-const contains =require('../models/contains');
