--- controllers/generalStatisticsController.js
@@ -1,1 +1,0 @@
-const generalStatistics =require('../models/generalStatistics');
