--- controllers/productDataController.js
@@ -1,1 +1,0 @@
-const productData =require('../models/productData');
