--- controllers/ratingController.js
@@ -1,1 +1,0 @@
-const rating = require('../models/rating');
