+++ models/User.js
@@ -9,5 +9,17 @@ const userSchema = new Schema({
     email:String,
     phoneNumber:Number,
     password:String,
+    likedProducts:[{
+        type:[mongoose.Types.ObjectId],
+        ref:'product'
+    }],
+    wishList:[{
+        type:mongoose.Types.ObjectId,
+        ref:'product'
+    }],
+    role:{
+        type:String,
+        default:'client'
+    }
 });
 module.exports = mongoose.model("user", userSchema);
