+++ models/seller.js
@@ -1,23 +1,14 @@
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 const sellerSchema = new Schema({
-    firstName:String,
-    lastName:String,
-    gender:String,
-    dateOfBirth:Date,
-    avatar:String,
-    email:String,
-    phoneNumber:Number,
-    password:String,
-    commerceRegistrNumber:Number,
-    age:Number,
-    isProffestionalAccount:{
-        type : Boolean,
-        default : false
+    userId:{
+        type:mongoose.Types.ObjectId,
+        ref:'user'
     },
     isAccepted:{
         type : Boolean,
         default : false
-    }
+    },
+    registerCommerce:String
 });
 module.exports = mongoose.model("seller", sellerSchema);
